snowleopard 64 bit warning death
make all errors from warnings when compiling on snowleopard die

as a note, the ukernel pid, and pthread self as id changes, probably
should be revisited, but for now i have changed nothing. it should
be a separate change.

Reviewed-on: http://gerrit.openafs.org/666
Tested-by: Derrick Brashear <shadow@dementia.org>

+++ src/volser/vsutils.c
@@ -425,7 +425,6 @@ vsu_ExtractName(char rname[], char name[])
     }
 }
 
-
 /* returns 0 if failed */
 afs_uint32
 vsu_GetVolumeID(char *astring, struct ubik_client *acstruct, afs_int32 *errp)
@@ -441,7 +440,7 @@ vsu_GetVolumeID(char *astring, struct ubik_client *acstruct, afs_int32 *errp)
 	char *end;
 	afs_uint32 result;
 	result = strtoul(astring, &end, 10);
-	if (result != ULONG_MAX && *end == '\0')
+	if (result != UINT_MAX && *end == '\0')
 	    return result;
     }
